requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai>=1.45.0
twilio==9.2.3
requests==2.32.3
pydantic==2.9.2
# LangChain + FAISS + PDF
langchain==0.3.4
langchain-community==0.3.3
faiss-cpu==1.8.0.post1
pypdf==4.3.1
reportlab==4.0.9
psycopg[binary]==3.2.1
